Modify AnswersDto JSON serialisation.
Move customisation up from AnswerDto to AnswersDto, in order to
eliminate the additional answers: { ... } wrapping.

We don't actually need to deserialise AnswersDto in production, but it's
*extremely* useful for testing.
