+++ src/main/kotlin/uk/gov/justice/digital/assessments/api/AnswerDto.kt
@@ -10,6 +10,11 @@ import com.fasterxml.jackson.databind.deser.std.StdDeserializer
 import com.fasterxml.jackson.databind.ser.std.StdSerializer
 import java.lang.IllegalStateException
 
+@JsonSerialize(using = AnswerDtoSerializer::class)
+data class AnswerDto(
+  val items: Collection<String>
+)
+
 object AnswerDtoSerializer : StdSerializer<AnswerDto>(AnswerDto::class.java) {
   override fun serialize(value: AnswerDto?, gen: JsonGenerator?, provider: SerializerProvider?) {
     if (value == null)
@@ -23,25 +28,3 @@ object AnswerDtoSerializer : StdSerializer<AnswerDto>(AnswerDto::class.java) {
   }
 }
 
-object AnswerDtoDeserializer : StdDeserializer<AnswerDto>(AnswerDto::class.java) {
-  override fun deserialize(p: JsonParser, ctxt: DeserializationContext): AnswerDto {
-    val node = ctxt.readTree(p)
-
-    if (node.isTextual) {
-      val item = ctxt.readValue(p, String::class.java)
-      return AnswerDto(listOf(item))
-    }
-    if (node.isArray) {
-      val items = node.elements().asSequence().map { it.asText() }.toList()
-      return AnswerDto(items)
-    }
-
-    throw IllegalStateException("Expected a string or array of strings to deserialise an Answer, but type was ${node.nodeType}")
-  }
-}
-
-@JsonSerialize(using = AnswerDtoSerializer::class)
-@JsonDeserialize(using = AnswerDtoDeserializer::class)
-data class AnswerDto(
-  val items: Collection<String>
-)
